Make color controller so we only mount once per map component
